# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: NextAPI Gateway
  version: 1.0.0
  summary: Unified B2B gateway for ByteDance Seedance 2.0 video generation.
  description: |
    Exposes a clean, OpenAI/Replicate/fal-style surface over ByteDance's
    Ark Seedance 2.0 / 2.0-fast / 1.x video models. All Ark-specific
    quirks (prompt-suffix flags, token-based pricing, 24h asset TTL,
    endpoint-id indirection) are hidden inside the gateway.

    Conventions:
      * Bearer auth. Business key (sk_live_* / sk_test_*) for /videos & /models;
        admin key (ak_live_* / ak_test_*) for /admin, /keys, /webhooks, /credits, /usage.
      * Idempotency-Key header on POSTs. 24h de-dup window.
      * All timestamps are RFC 3339 UTC unless stated otherwise.
      * Cursor pagination on list endpoints.
      * Webhook signatures use HMAC-SHA256 over `t.` + raw body.
  contact:
    name: NextAPI Platform
    url: https://docs.nextapi.top
  license:
    name: Proprietary
servers:
  - url: https://api.nextapi.top/v1
    description: Production
  - url: https://api.sandbox.nextapi.top/v1
    description: Sandbox (sk_test_* keys, synthetic output)
security:
  - bearerBusiness: []
...
paths:
...
  /models:
    get:
      tags: [models]
      operationId: listModels
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ModelPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /models/{model_id}:
    get:
      tags: [models]
      operationId: getModel
      parameters:
        - name: model_id
          in: path
          required: true
          schema: { type: string, example: seedance-2.0 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Model' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /billing/ledger:
    get:
      tags: [credits]
      operationId: listLedger
      security: [ { bearerAdmin: [] } ]
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [data, has_more]
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/LedgerEntry' } }
                  has_more: { type: boolean }
...
components:
  securitySchemes:
    bearerBusiness:
      type: http
      scheme: bearer
      bearerFormat: sk_{live|test}_...
    bearerAdmin:
      type: http
      scheme: bearer
      bearerFormat: ak_{live|test}_...
  parameters:
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
...
  responses:
    Unauthorized:
      description: Missing or invalid API key.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: invalid_request.resolution_unsupported }
            message: { type: string }
            request_id: { type: string }
            param: { type: string }
            doc_url: { type: string, format: uri }
...
    AspectRatio:
      type: string
      enum: ["16:9", "4:3", "1:1", "3:4", "9:16", "21:9"]
    Resolution:
      type: string
      enum: [480p, 720p, 1080p]
...
    Model:
      type: object
      required: [id, family, max_duration_seconds, supported_resolutions]
      properties:
        id: { type: string }
        family: { type: string }
        description: { type: string }
        modality_support:
          type: object
          properties:
            text_to_video: { type: boolean }
            image_to_video: { type: boolean }
            video_to_video: { type: boolean }
            audio_to_video: { type: boolean }
            multimodal_reference: { type: boolean }
        max_duration_seconds: { type: integer }
        min_duration_seconds: { type: integer, default: 4 }
        supports_auto_duration: { type: boolean }
        supported_resolutions:
          type: array
          items: { $ref: '#/components/schemas/Resolution' }
        supported_aspect_ratios:
          type: array
          items: { $ref: '#/components/schemas/AspectRatio' }
        supports_audio_output: { type: boolean }
        price_cents_per_second:
          type: object
          additionalProperties: { type: integer }
        status: { type: string, enum: [ga, beta, deprecated] }
    ModelPage:
      type: object
      required: [data, has_more]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Model' }
        has_more: { type: boolean }
        next_cursor: { type: string, nullable: true }
...
    LedgerEntry:
      type: object
      properties:
        id: { type: integer, format: int64 }
        org_id: { type: string }
        delta_credits: { type: integer }
        delta_cents: { type: integer, nullable: true }
        reason: { type: string, enum: [signup_bonus, topup, consumption, reservation, reconciliation, refund, adjustment] }
        job_id: { type: string, nullable: true }
        note: { type: string }
        created_at: { type: string, format: date-time }
...